move tests from src/components/common folder to test folder/components/common
