+++ frontend/test/components/common/DeleteModal.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react";
 import { vi } from "vitest";
 import userEvent from "@testing-library/user-event";
 import "@testing-library/jest-dom";
-import DeleteModal from "./DeleteModal";
+import DeleteModal from "../../../src/components/common/DeleteModal";
 
 describe("DeleteModal", () => {
   it("renders modal if open", () => {
